requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydub
python-dotenv
python-multipart
open-interpreter
pandas
pyspark
seaborn
matplotlib